bio_auth_backend/backend/requirements.txt
Flask==2.3.3
python-dotenv
face_recognition
opencv-python
numpy
requests
qrcode
cryptography
pillow